src/test/java/com/quinbay/reactivepractice/serviceimpl/BookServiceTest.java
package com.quinbay.reactivepractice.serviceimpl;

import org.junit.jupiter.api.Test;
import reactor.test.StepVerifier;

import static org.junit.jupiter.api.Assertions.*;

class BookServiceTest {

    private BookInfoService bookInfoService = new BookInfoService();
    private ReviewService reviewService = new ReviewService();
    private BookService bookService = new BookService(bookInfoService, reviewService);

    @Test
    void getBooks() {
        var books = bookService.getBooks();
        StepVerifier.create(books)
                .assertNext(book -> {
                    assertEquals("B1", book.getBookInfo().getTitle());
                    assertEquals(2, book.getReviews().size());
                })
                .assertNext(book -> {
                    assertEquals("B2", book.getBookInfo().getTitle());
                    assertEquals(2, book.getReviews().size());
                })
                .assertNext(book -> {
                    assertEquals("B3", book.getBookInfo().getTitle());
                    assertEquals(2, book.getReviews().size());
                })
                .verifyComplete();
    }


    @Test
    void getBookById() {
        var book = bookService.getBookById(1);
        StepVerifier.create(book)
                .assertNext(book1 -> {
                    assertEquals("B1", book1.getBookInfo().getTitle());
                    assertEquals(2, book1.getReviews().size());
                })
                .verifyComplete();
    }
}